Add validation to POST/PUT method, Implementing Generic ExceptionHandler

+++ src/main/java/com/gilad/employee_management/controller/EmployeeController.java
@@ -6,7 +6,10 @@ import com.gilad.employee_management.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import javax.validation.Valid;
+import java.net.URI;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -19,38 +22,42 @@ public class EmployeeController {
     @Autowired
     private EmployeeRepository employeeRepository;
 
-     @GetMapping("/employees")
-    public List<Employee> getAllEmployees(){
-         return employeeRepository.findAll();
-     }
+    @GetMapping("/employees")
+    public List<Employee> getAllEmployees() {
+        return employeeRepository.findAll();
+    }
 
-     @PostMapping("/employees")
-     public Employee createEmployee (@RequestBody Employee employee){
-         return employeeRepository.save(employee);
-     }
+    @PostMapping("/employees")
+    public ResponseEntity<Employee> createEmployee(@Valid @RequestBody Employee employee) {
+        return ResponseEntity.ok(employeeRepository.save(employeeRepository.save(employee)));
+    }
 
-     @GetMapping("/employees/{id}")
-    public ResponseEntity<Employee> getEmployeeById(@PathVariable Long id){
-         Employee employee =  employeeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee with the id - "+id+" not exist"));
-         return ResponseEntity.ok(employee);
-     }
+    @GetMapping("/employees/{id}")
+    public ResponseEntity<Employee> getEmployeeById(@PathVariable Long id) {
+        Employee employee = employeeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee with the id - " + id + " not exist"));
+        if (employee == null)
+            throw new ResourceNotFoundException("id-" + id);
 
-     @PutMapping("/employees/{id}")
-     public ResponseEntity<Employee> updateEmployee(@PathVariable Long id, @RequestBody Employee employeeDetails){
-         Employee employee =  employeeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee with the id - "+id+" not exist"));
-         employee.setFirstName(employeeDetails.getFirstName());
-         employee.setLastName(employeeDetails.getLastName());
-         employee.setEmail(employeeDetails.getEmail());
-         return ResponseEntity.ok(employeeRepository.save(employee));
+        return ResponseEntity.ok(employee);
+    }
 
-     }
+    @PutMapping("/employees/{id}")
+    public ResponseEntity<Employee> updateEmployee(@PathVariable Long id, @Valid @RequestBody Employee employeeDetails) {
+        Employee employee = employeeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee with the id - " + id + " not exist"));
+        employee.setFirstName(employeeDetails.getFirstName());
+        employee.setLastName(employeeDetails.getLastName());
+        employee.setEmail(employeeDetails.getEmail());
+        return ResponseEntity.ok(employeeRepository.save(employee));
+    }
 
-     @DeleteMapping("/employees/{id}")
-    public ResponseEntity<Map<String,Boolean>> deleteEmployee(@PathVariable Long id){
-         Employee employee =  employeeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee with the id - "+id+" not exist"));
-         employeeRepository.delete(employee);
-         Map <String,Boolean> res = new HashMap<>();
-         res.put("delete",Boolean.TRUE);
-         return ResponseEntity.ok(res);
-     }
+    @DeleteMapping("/employees/{id}")
+    public ResponseEntity<Map<String, Boolean>> deleteEmployee(@PathVariable Long id) {
+        Employee employee = employeeRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee with the id - " + id + " not exist"));
+        if (employee == null)
+            throw new ResourceNotFoundException("id-" + id);
+        employeeRepository.delete(employee);
+        Map<String, Boolean> res = new HashMap<>();
+        res.put("delete", Boolean.TRUE);
+        return ResponseEntity.ok(res);
+    }
 }
